requirements.txt
flask
flask-sqlalchemy
flask-smorest
python-dotenv
marshmallow
psycopg2-binary
passlib
Flask-JWT-Extended
flask-migrate
gunicorn
psycopg2
requests
redis
rq
Flask-RESTful